+++ SolBo/SolBo.Shared/Services/Implementations/MarketService.cs
@@ -4,7 +4,7 @@ namespace SolBo.Shared.Services.Implementations
 {
     public class MarketService : IMarketService
     {
-        public FundResponse AvailableQuote(int fundPercentage, decimal availableQuote, int precision)
+        public FundResponse AvailableQuote(decimal fundPercentage, decimal availableQuote, int precision)
         {
             var result = new FundResponse();
 
@@ -17,7 +17,7 @@ namespace SolBo.Shared.Services.Implementations
 
             return result;
         }
-        public MarketResponse IsGoodToBuy(int percentPriceDrop, decimal storedPriceAverage, decimal currentPrice)
+        public MarketResponse IsGoodToBuy(decimal percentPriceDrop, decimal storedPriceAverage, decimal currentPrice)
         {
             return new MarketResponse
             {
@@ -27,7 +27,7 @@ namespace SolBo.Shared.Services.Implementations
                 PercentChanged = decimal.Round(100 - (currentPrice / storedPriceAverage * 100), 2)
             };
         }
-        public MarketResponse IsGoodToSell(int percentPriceRise, decimal storedPriceAverage, decimal currentPrice)
+        public MarketResponse IsGoodToSell(decimal percentPriceRise, decimal storedPriceAverage, decimal currentPrice)
         {
             return new MarketResponse
             {
@@ -37,7 +37,7 @@ namespace SolBo.Shared.Services.Implementations
                 PercentChanged = decimal.Round(100 - (storedPriceAverage * 100 / currentPrice), 2)
             };
         }
-        public MarketResponse IsStopLossReached(int percentStopLoss, decimal storedPriceAverage, decimal currentPrice)
+        public MarketResponse IsStopLossReached(decimal percentStopLoss, decimal storedPriceAverage, decimal currentPrice)
         {
             return new MarketResponse
             {
